Fix blur event interfering with selecting options
When we try to click on one of the menu options, the onblur
event fires first, removing the menu items from under our nose.
This messes up the onclick and makes so it never ends up happening.

The workaround was an ugly settimeout. I need to find out a better way.